+++ spec/truffle/rubygems/default_gems_list_spec.rb
@@ -10,8 +10,12 @@
 require_relative '../../ruby/spec_helper'
 require 'rubygems'
 
-describe "Default gem specs" do
-  it "@path_to_default_spec_map should be non-empty" do
-    Gem.instance_variable_get(:@path_to_default_spec_map).empty?.should == false
+describe "Truffle::GemUtil::DEFAULT_GEMS" do
+  it "is a Hash listing all default gem paths" do
+    hash = Dir.children("#{Gem.default_dir}/specifications/default").sort.map do |spec|
+      spec.split("-").first # 'io' for gem 'io-console' required as 'io/console'
+    end.to_h { |name| [name, true] }
+
+    Truffle::GemUtil::DEFAULT_GEMS.should == hash
   end
 end