+++ tool/jt.rb
@@ -29,7 +29,7 @@ TRUFFLERUBY_DIR = File.expand_path('../..', File.realpath(__FILE__))
 GRAAL_DIR = File.expand_path('../graal', TRUFFLERUBY_DIR)
 PROFILES_DIR = "#{TRUFFLERUBY_DIR}/profiles"
 
-TRUFFLERUBY_GEM_TEST_PACK_VERSION = '42919151f2aed1270d080c9eba78e4cb017feec1'
+TRUFFLERUBY_GEM_TEST_PACK_VERSION = '4b32d86150e5dfd4b25f67e96ab7c7f15424d3f7'
 
 JDEBUG = '--vm.agentlib:jdwp=transport=dt_socket,server=y,address=8000,suspend=y'
 METRICS_REPS = Integer(ENV['TRUFFLERUBY_METRICS_REPS'] || 10)